BiocUpgrade {BiocInstaller}	R Documentation

Upgrade Bioconductor to the latest version available for this version of R

Description

Downloads the latest version of the BiocInstaller package, and upgrades all currently installed packages to the latest repositories for this version of R.

To upgrade, use:


    ## 'http' if 'https' unavailable
    source("https://bioconductor.org/biocLite.R")
    biocLite("BiocUpgrade")
  

See Also

biocLite Installs/updates Bioconductor/CRAN packages.

chooseBioCmirror lets you choose from a list of all public Bioconductor mirror URLs.

chooseCRANmirror lets you choose from a list of all public CRAN mirror URLs.

biocinstallRepos returns the Bioconductor and CRAN repositories used by biocLite.

install.packages installs the packages themselves.

Examples


## Not run: 
source("https://bioconductor.org/biocLite.R")
biocLite("BiocUpgrade")

## End(Not run)

[Package BiocInstaller version 1.30.0 Index]