biocLite {BiocInstaller}	R Documentation

Install or update Bioconductor and CRAN packages

Description

biocLite installs or updates Bioconductor and CRAN packages in a Bioconductor release. Upgrading to a new Bioconductor release requires additional steps; see https://bioconductor.org/install.

BIOCINSTALLER_ONLINE_DCF is an environment variable or global options() which, when set to FALSE, uses configuration information from a local archive rather than consulting the current online version.

Usage


biocLite(pkgs=c("Biobase", "IRanges", "AnnotationDbi"),  
    suppressUpdates=FALSE,
    suppressAutoUpdate=FALSE,
    siteRepos=character(),
    ask=TRUE, ...)

Arguments

pkgs
character() of package names to install or update. A missing value and suppressUpdates=FALSE updates installed packages, perhaps also installing Biobase, IRanges, and AnnotationDbi if they are not already installed. Package names containing a ‘/’ are treated as github repositories and installed using the install_github() function of the remotes package.

suppressUpdates
logical(1) or character(). When FALSE, biocLite asks the user whether old packages should be update. When TRUE, the user is not prompted to update old packages. When character() a vector specifying which packages to NOT update.

suppressAutoUpdate
logical(1) indicating whether the BiocInstaller package updates itself.

siteRepos
character() representing an additional repository in which to look for packages to install. This repository will be prepended to the default repositories (which you can see with biocinstallRepos).

ask
logical(1) indicating whether to prompt user before installed packages are updated, or the character string 'graphics', which brings up a widget for choosing which packages to update. If TRUE, user can choose whether to update all outdated packages without further prompting, to pick and choose packages to update, or to cancel updating (in a non-interactive session, no packages will be updated). Otherwise, the value is passed to update.packages.

...
Additional arguments.

When installing CRAN or Bioconductor packages, typical arguments include: lib.loc, passed to old.packages and used to determine the library location of installed packages to be updated; and lib, passed to install.packages to determine the library location where pkgs are to be installed.

When installing github packages, ... is passed to the remotes package functions install_github and install. A typical use is to build vignettes, via dependencies=TRUE, build_vignettes=TRUE.

Details

Installation of Bioconductor and CRAN packages use R's standard functions for library management – install.packages(), available.packages(), update.packages(). Installation of github packages uses the install_github() function from the remotes package. For this reason it usually makes sense, when complicated installation options are needed, to invoke biocLite() separately for Bioconductor / CRAN packages and for github packages.

Setting BIOCINSTALLER_ONLINE_DCF to FALSE can speed package loading when internet access is slow or non-existent, but may result in out-of-date information about the current release and development versions of Bioconductor.

Value

biocLite() returns the pkgs argument, invisibly.

See Also

biocinstallRepos returns the Bioconductor and CRAN repositories used by biocLite.

install.packages installs the packages themselves.

update.packages updates all installed packages.

chooseBioCmirror lets you choose from a list of all public Bioconductor mirror URLs.

chooseCRANmirror lets you choose from a list of all public CRAN mirror URLs.

monograph_group, RBioinf_group and biocases_group return package names associated with Bioconductor publications.

all_group returns the names of all Bioconductor software packages.

Examples


## Not run: 
## Change default Bioconductor and CRAN mirrors
chooseBioCmirror()
chooseCRANmirror()

## If you don't have the BiocInstaller package installed, you can 
## quickly install and load it as follows:
source("https://bioconductor.org/biocLite.R")  # 'http' if 'https' unavailable

## The most recent version of the BiocInstaller package is now loaded.
## No need to load it with library().

## installs default packages (if not already installed) and updates
## previously installed packages
biocLite()

## Now install a CRAN package:
biocLite("survival")

## install a Bioconductor package, but don't update all installed
## packages as well:
biocLite("GenomicRanges", suppressUpdates=TRUE)

## Install default packages, but do not update any package whose name
## starts with "org." or "BSgenome."
biocLite(suppressUpdates=c("^org\.", "^BSgenome\."))

## install a package from source:
biocLite("IRanges", type="source")

## install all Bioconductor software packages
biocLite(all_group())


## End(Not run)
## Show the Bioconductor and CRAN repositories that will be used to
## install/update packages.
biocinstallRepos()

## Use local archive rather than current online configuration
## information. Set this prior to loading the BiocInstaller package.
options(BIOCINSTALLER_ONLINE_DCF = FALSE)

[Package BiocInstaller version 1.30.0 Index]