biocUpdatePackages {BiocInstaller}	R Documentation

Update previously installed Bioconductor or CRAN packages and their dependencies.

Description

Update previously installed Bioconductor and CRAN packages and their dependencies. Use biocLite to install new packages or to update all out-of-date packages. Upgrading to a new Bioconductor release requires additional steps; see https://bioconductor.org/install.

Usage


biocUpdatePackages(pkgs, dependencies = NA, repos=biocinstallRepos(), ...)

Arguments

pkgs
character() of package names to install or update.

dependencies
character() describing out-of-date dependencies that are also updated. Defaults to c("Depends", "Imports", "LinkingTo") but can be a subset of c("Depends", "Imports", "LinkingTo", "Suggests", "Enhances").

repos
character() of named repositories in which to look for package updates, in the style of biocinstallRepos().

...
Additional arguments, passed to update.packages. For example, ask=FALSE to avoid prompts to update individual packages.

Value

‘NULL’, invisibly.

Author(s)

Martin Morgan mtmorgan@fhcrc.org

See Also

biocLite

Examples


## Not run: 
biocUpdatePackages("GenomicRanges", ask=FALSE)

## End(Not run)

[Package BiocInstaller version 1.30.0 Index]