biocValid {BiocInstaller}	R Documentation

Validate installed package versions against biocLite versions.

Description

Check that installed packages are consistent (neither out-of-date nor too new) with the version of R and Bioconductor in use, using biocLite for validation.

Usage


biocValid(pkgs = installed.packages(lib.loc, priority = priority),
    lib.loc = NULL, priority = "NA", type = getOption("pkgType"),
    filters = NULL, silent = FALSE, ..., fix=FALSE)

Arguments

pkgs
A character list of package names for checking, or a matrix as returned by installed.packages.

lib.loc
The library location(s) of packages to be validated; see installed.packages.

priority
check validity of all, “base”, or “recommended” packages; see installed.packages.

type
The type of available package (e.g., binary, source) to check validity against; see available.packages.

filters
Filter available packages to check validity against; see available.packages.

silent
Report how packages are invalid (silent=FALSE, default) and abort execution, or return a logical(1) (silent=TRUE) indicating the overall validity of installed packages.

...
Additional arguments, passed to biocLite when fix=TRUE.

fix
When TRUE, invoke biocLite to reinstall (update or downgrade, as appropriate) invalid packages.

Details

This function compares the version of installed packages to the version of packages associated with the version of R and Bioconductor appropriate for the BiocInstaller package currently in use.

Packages are reported as ‘out-of-date’ if a more recent version is available at the repositories specified by biocinstallRepos(). Usually, biocLite() is sufficient to update packages to their most recent version.

Packages are reported as ‘too new’ if the installed version is more recent than the most recent available in the biocinstallRepos() repositories. It is possible to down-grade by re-installing a too new package “PkgA” with biocLite("PkgA"). It is important for the user to understand how their installation became too new, and to avoid this in the future.

Value

logical(1) indicating overall validity of installed packages.

Author(s)

Martin Morgan mtmorgan@fhcrc.org

See Also

biocLite to update installed packages.

Examples

try(biocValid())

[Package BiocInstaller version 1.30.0 Index]