biocinstallRepos {BiocInstaller}	R Documentation

Display current Bioconductor and CRAN repositories.

Description

Displays the URLs of the repositories used by biocLite to install Bioconductor and CRAN packages.

Usage


biocinstallRepos(siteRepos=character(), version=biocVersion())

Arguments

siteRepos
(Optional) character(1) representing an additional repository (e.g., a URL to an organization's internally maintained repository) in which to look for packages to install. This repository will be prepended to the default repositories returned by the function.

version
(Optional) character(1) or package_version indicating the Bioconductor version (e.g., “3.1”) for which repositories are required.

Value

Named character() of repositories.

See Also

biocLite Installs/updates Bioconductor/CRAN packages.

install.packages installs the packages themselves.

chooseBioCmirror lets you choose from a list of all public Bioconductor mirror URLs.

chooseCRANmirror lets you choose from a list of all public CRAN mirror URLs.

Examples


biocinstallRepos()

## Choose mirrors
## Not run: 
chooseCRANmirror()
chooseBioCmirror()

## End(Not run)

[Package BiocInstaller version 1.30.0 Index]