Package Groups {BiocInstaller}	R Documentation

Convenience functions to return package names associated with Bioconductor publications.

Description

Returns character vectors of packages associated with Bioconductor publications, which can then be passed to biocLite()

Usage


monograph_group()
RBioinf_group()
biocases_group()
all_group()

Value

character() of package names.

See Also

biocLite Installs/updates Bioconductor/CRAN packages.

biocinstallRepos returns the Bioconductor and CRAN repositories used by biocLite.

install.packages installs the packages themselves.

chooseBioCmirror lets you choose from a list of all public Bioconductor mirror URLs.

chooseCRANmirror lets you choose from a list of all public CRAN mirror URLs.

Examples



## Get the names of packages used in the book
## "Bioconductor Case Studies":
biocases_group()

## Get the names of packages used in the book
## "R Programming for Bioinformatics":
RBioinf_group()

## Get the names of packages used in the monograph
## "Bioinformatics and Computational Biology Solutions
## Using R and Bioconductor":
monograph_group()

## Get the names of all Bioconductor software packages
all_group()

[Package BiocInstaller version 1.30.0 Index]