biocLite {BiocInstaller}	R Documentation

Install or update Bioconductor and CRAN packages

Description

biocLite installs or updates Bioconductor and CRAN packages, ensuring that packages from the appropriate version of Bioconductor are installed, and that all packages remain up to date.

Usage

biocLite (pkgs=c("Biobase","IRanges","AnnotationDbi"),  
  suppressUpdates=FALSE,
  suppressAutoUpdate=FALSE,
  ask=TRUE, ...)

Arguments

pkgs
character() of package names to install or update. A value of character(0) and suppressUpdates=FALSE updates packages without installing new ones.

suppressUpdates
logical(1) indicating whether to suppress automatic updating of all installed packages, or character() of regular expressions specifying which packages to NOT automatically update.

suppressAutoUpdate
logical(1) indicating whether the BiocInstaller package updates itself.

ask
logical(1) indicating whether to prompt user before installed packages are updated, or the character string 'graphics', which brings up a widget for choosing which packages to update. If TRUE, user can choose whether to update all outdated packages without further prompting, to pick and choose packages to update, or to cancel updating (in a non-interactive session, no packages will be updated). Otherwise, the value is passed to update.packages.

...
Additional arguments passed to install.packages.

Details

The biocLite() function is generally used after sourcing the file biocLite.R. This will install the BiocInstaller package if it is not already installed. See example section for more information.

Value

biocLite() returns the pkgs argument, invisibly.

See Also

biocinstallRepos returns the Bioconductor and CRAN repositories used by biocLite.

install.packages installs the packages themselves.

update.packages updates all installed packages.

chooseBioCmirror lets you choose from a list of all public Bioconductor mirror URLs.

chooseCRANmirror lets you choose from a list of all public CRAN mirror URLs.

monograph_group, RBioinf_group and biocases_group return package names associated with Bioconductor publications.

all_group returns the names of all Bioconductor software packages.

Examples

## Not run: 
## Change default Bioconductor and CRAN mirrors
chooseBioCmirror()
chooseCRANmirror()

## If you don't have the BiocInstaller package installed, you can 
## quickly install and load it as follows:
source("http://bioconductor.org/biocLite.R")

## The most recent version of the BiocInstaller package is now loaded.
## No need to load it with library().
biocLite() # installs default packages

## Now install a CRAN package:
biocLite("survival")

## install a Bioconductor package, but don't update all installed
## packages as well:
biocLite("GenomicRanges", suppressUpdates=TRUE)

## Install default packages, but do not update any package whose name
## starts with "org." or "BSgenome."
biocLite(suppressUpdates=c("^org\.", "^BSgenome\."))

## install a package from source:
biocLite("IRanges", type="source")

## don't install any packages, but update installed packages
biocLite(character(0), ask=FALSE)

## install all Bioconductor software packages
biocLite(all_group())

## End(Not run)
## Show the Bioconductor and CRAN repositories that will be used to
## install/update packages.
biocinstallRepos()

[Package BiocInstaller version 1.2.1 Index]