biocinstallRepos {BiocInstaller}	R Documentation

Display current Bioconductor and CRAN repositories.

Description

Displays the URLs of the repositories used by biocLite to install Bioconductor and CRAN packages.

Usage

biocinstallRepos()

Value

Named character() of repositories.

See Also

biocLite Installs/updates Bioconductor/CRAN packages.

install.packages installs the packages themselves.

chooseBioCmirror lets you choose from a list of all public Bioconductor mirror URLs.

chooseCRANmirror lets you choose from a list of all public CRAN mirror URLs.

Examples

biocinstallRepos()

## Choose mirrors
## Not run: 
chooseCRANmirror()
chooseBioCmirror()

## End(Not run)

[Package BiocInstaller version 1.2.1 Index]